+++ package.json
@@ -17,7 +17,8 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
     "react-markdown": "^8.0.3",
-    "react-router-dom": "^6.4.0"
+    "react-router-dom": "^6.4.2",
+    "use-media": "^1.4.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.17",
