+++ src/components/Layout/Container.module.css
@@ -1,0 +1,5 @@
+.container{
+  width: min(50rem, 100%);
+  flex-wrap: wrap;
+  background-color: brown;
+}
