+++ src/components/Layout/Footer/footer.module.css
@@ -1,30 +1,32 @@
-/* footer .gradient {
+footer .gradient {
   height: 8rem;
+  width: 100%;
   background: linear-gradient(to bottom, rgb(38, 38, 38), rgb(18, 18, 18));
-} */
+}
 
 footer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
   width: 100%;
-  height: 40.3rem;
-  padding-inline: 9rem;
-  display: flex;
-  align-items: inherit;
-  justify-content: space-around;
   background-color: rgba(0, 0, 0, 0.6);
 }
+
+.flex{
+  display: flex;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  margin: 0 auto;
+  padding-top: 7.2rem;
+  padding-inline: 9rem;
+  /* background-color: blue; */
+}
+
 footer .containerLogoAndMessage {
   display: flex;
   align-items: center;
   justify-content: center;
   flex-direction: column;
   height: 8.5rem;
-  margin-top: 5rem;
 }
 footer .containerLogoAndMessage .guinaRangerLogo img {
-  width: 43rem;
   height: 6rem;
   max-width: 100%;
 }
@@ -37,7 +39,8 @@ footer .containerLogoAndMessage .fraseLogo {
 }
 footer .containerSocialLinks {
   text-align: center;
-  margin-top: 22.5rem;
+  margin-top: 20rem;
+  padding-bottom: 5rem;
   padding-inline: 2rem;
 }
 footer .containerSocialLinks .redeSociaisText {
@@ -50,7 +53,7 @@ footer .containerSocialLinks ul {
   align-items: center;
   justify-content: center;
   align-items: left;
-  gap: 6.5rem;
+  gap: 5rem;
   margin-block: 4rem;
 }
 footer .containerSocialLinks ul li a {
@@ -65,37 +68,51 @@ footer .containerSocialLinks ul li:hover {
   color: rgb(255, 231, 17);
 }
 footer .containerSocialLinks .copyright {
-  margin-top: 5rem;
+  margin-top: 1rem;
   font-weight: 400;
   font-size: 1.4rem;
   line-height: 1.7rem;
   font-style: italic;
 }
-footer .containerMenuFooter {
+footer .containerListFooter {
   height: 12rem;
   padding: 1rem;
-  margin-top: 7rem;
   text-align: center;
 }
-footer .containerMenuFooter .containeruls {
+footer .containerListFooter .list {
   display: flex;
   align-items: center;
   justify-content: center;
   gap: 7rem;
 }
-footer .containerMenuFooter .containeruls ul {
+footer .containerListFooter .list ul {
   display: flex;
   flex-direction: column;
   gap: 0.7rem;
 }
-footer .containerMenuFooter .containeruls ul p {
+footer .containerListFooter .list ul p {
   margin-bottom: 0.7rem;
   font-size: 1.8rem;
   font-weight: 600;
   line-height: 2.2rem;
 }
-footer .containerMenuFooter .containeruls ul li {
+footer .containerListFooter .list ul li {
   font-size: 1.4rem;
   font-weight: 400;
   line-height: 1.7rem;
-}/*# sourceMappingURL=footer.module.css.map */
+}
+
+
+@media (max-width: 1600px){
+  .containerListFooter{
+    display: none;
+    background-color: red;
+  }
+  .flex{
+    flex-direction: column;
+    padding-top: 5rem;
+  }
+  footer .containerSocialLinks{
+    margin-top: 10rem;
+  }
+}
