+++ src/components/Layout/Menu/isDesktop/Menu.module.css
@@ -69,5 +69,3 @@
   filter: brightness(0.8);
   background-color: rgb(255, 255, 255);
 }
-
-/*navigation*//*# sourceMappingURL=menu.module.css.map */
